requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-api-python-client==2.121.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pydantic==2.6.4
